Rename classes after review.
